feat: 존재하지 않는 역 삭제 예외처리

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -8,6 +8,7 @@ import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 import wooteco.subway.station.exception.DuplicatedStationNameException;
+import wooteco.subway.station.exception.NoSuchStationException;
 
 @Service
 public class StationService {
@@ -39,6 +40,9 @@ public class StationService {
     }
 
     public void deleteStationById(Long id) {
-        stationDao.deleteById(id);
+        if (!stationDao.isHaveStationById(id)) {
+            stationDao.deleteById(id);
+        }
+        throw new NoSuchStationException("존재하지 않는 역은 삭제 할 수 없습니다.");
     }
 }
